requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
google-generativeai
celery[redis]
redis
PyMuPDF
chromadb